Moved input reader and output write to agogosml
