+++ agogosml/Pipfile
@@ -13,6 +13,7 @@ pytest = "==3.4.2"
 pytest-runner = "==2.11.1"
 Sphinx = "==1.7.1"
 twine = "==1.10.0"
+pylint = "*"
 
 [packages]
 confluent-kafka = "*"
