+++ agogosml/agogosml/writer/output_writer_factory.py
@@ -6,7 +6,7 @@ import os
 from agogosml.streaming_client.flask_http_listener_client import FlaskHttpListenerClient
 from agogosml.streaming_client.eventhub_streaming_client import EventHubStreamingClient
 from agogosml.streaming_client.kafka_streaming_client import KafkaStreamingClient
-from output_writer import OutputWriter
+from .output_writer import OutputWriter
 
 
 class OutputWriterFactory:
