+++ input_reader/Dockerfile.input_reader
@@ -5,11 +5,6 @@ FROM ${CONTAINER_REG}agogosml/agogosml:${AGOGOSML_TAG} as builder
 
 WORKDIR /usr/src/agogosml
 COPY ./input_reader ./input_reader
-COPY ./tests/ ./tests
-
-#RUN pip install pytest pytest-runner pylint && \
-#    pylint ./input_reader/*.py && \
-#    pytest ./tests/*.py
 
 # Release
 FROM ${CONTAINER_REG}agogosml/agogosml:${AGOGOSML_TAG} AS input_reader
