+++ output_writer/Dockerfile.output_writer
@@ -6,11 +6,6 @@ FROM ${CONTAINER_REG}agogosml/agogosml:${AGOGOSML_TAG} as builder
 # Copy the app
 WORKDIR /usr/src/agogosml
 COPY ./output_writer ./output_writer
-COPY ./tests/ ./tests
-
-#RUN pip install pytest pytest-runner pylint && \
-#    pylint ./output_writer/*.py && \
-#    pytest ./tests/*.py
 
 # Release
 FROM ${CONTAINER_REG}agogosml/agogosml:${AGOGOSML_TAG} AS output_writer
